refactor cast node removal
